legacy: move pbkdf2 static method from FlexiKey to Pbkdf2FileMixin, refs #9556
